requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
sqlite3

# 向量数据库
chromadb==0.4.18

# LangChain相关
langchain==0.0.340
langchain-community==0.0.1
langchain-openai==0.0.2

# 文档处理
pypdf2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
markdown==3.5.1

# 机器学习
numpy==1.24.3
scikit-learn==1.3.2

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 配置管理
pyyaml==6.0.1
python-dotenv==1.0.0

# 日志和监控
loguru==0.7.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 类型提示
typing-extensions==4.8.0

# 数据验证
pydantic==2.5.0